+++ src/Contracts/NotificationHandlerContract.php
@@ -13,6 +13,8 @@ interface NotificationHandlerContract
 {
     /**
      * Executes the handler.
+     *
+     * @psalm-suppress MissingReturnType - @todo add return type
      */
     public function execute(): void;
 }