+++ src/Events/AppStore/EventFactory.php
@@ -21,7 +21,7 @@ class EventFactory
     public static function create(ServerNotificationContract $notification): PurchaseEventContract
     {
         $type = $notification->getType();
-        $className = __NAMESPACE__.'\\'.ucfirst(Str::camel(strtolower($type)));
+        $className = "\Imdhemy\Purchases\Events\AppStore\\".ucfirst(Str::camel(strtolower($type)));
         assert(class_exists($className) && is_subclass_of($className, PurchaseEventContract::class));
 
         return new $className($notification);